MG: navigation - router
