+++ Cinema/qac/src/Components/Listings/Listings.jsx
@@ -1,6 +1,11 @@
+import Navbar from "../Navbar/Navbar";
+
 const Listings = () =>{
     return(
+        <>
+        <Navbar/>
         <h1>hello</h1>
+        </>
     )
 }
