--- Cinema/qac/src/Components/Navbar/Nav.jsx
@@ -1,10 +1,0 @@
-const Nav = () => {
-    return( 
-        <>
-            <Link to="/">Home</Link>
-            <Link to="/listings">Listings</Link>
-        </>
-    )
-}
-
-export default Nav;
